Se separo el modulo main a ReadInstance.py, el cual se encarga de leer el archivo de instancias y extrea toda la informacion necesaria. Tambien se creo la carpeta Instances para poder colocar y probar distintas instancias
